requirements.txt
langchain==0.1.5
langchain_experimental==0.0.50
langchain_openai==0.0.5
pandas==2.0.3
python-dotenv==1.0.1
streamlit==1.30.0
tabulate==0.9.0
